trading-platform-ui/requirements.txt

# Flask web framework
Flask

# Database ORM
Flask-SQLAlchemy

# Database migrations
Flask-Migrate

# Environment variables management
python-dotenv

# PostgreSQL database adapter
psycopg2-binary

requests

